# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DOCX Exporter REST API",
    "version": "v1"
  },
  "paths": {
...
    "/api/configuration/default-settings": {
      "get": {
        "operationId": "checkDefaultSettings",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "default": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationStatus"
                }
              }
            },
            "description": "Successfully retrieved default settings configuration"
          }
        },
        "summary": "Checks default settings configuration"
      }
    },
...
    "/api/configuration/document-properties-pane-config": {
      "get": {
        "operationId": "checkDocumentPropertiesPaneConfig",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "schema": {
              "default": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationStatus"
                }
              }
            },
            "description": "Successfully retrieved document properties pane configuration"
          }
        },
        "summary": "Checks document properties pane configuration"
      }
    },
...
    "/api/projects/{projectId}/collections/{collectionId}/documents": {
      "get": {
        "operationId": "getDocumentsFromCollection",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Collection ID",
            "in": "path",
            "name": "collectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The specific revision of the provided collection",
            "in": "query",
            "name": "revision",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentCollectionEntry"
                }
              }
            },
            "description": "Successfully retrieved the list of documents in the collection"
          }
        },
        "summary": "Get documents from collection",
        "tags": [
          "Collections"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ConfigurationStatus": {
        "properties": {
          "details": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "enum": [
              "OK",
              "WARNING",
              "ERROR"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "DocumentCollectionEntry": {
        "description": "Details about document from collection",
        "properties": {
          "documentName": {
            "description": "The name of the document",
            "type": "string"
          },
          "projectId": {
            "description": "The unique identifier for the project",
            "example": "elibrary",
            "type": "string"
          },
          "revision": {
            "description": "The revision of the document",
            "type": "string"
          },
          "spaceId": {
            "description": "The unique identifier of the space within the project",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
...
}